requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.8.0
yt-dlp>=2023.0.0
youtube-search-python>=1.4.9
spotipy>=2.23.0
lyricsgenius>=4.12.0
requests>=2.31.0
Pillow>=10.0.0